Fixed PermitsQuery and ForbidsQuery: building permits/forbids config based on the other (base) config.
